Split the factory and disable abstract factory for now.
